[fix] Cast learning-to-rank loss logits to float32 for bf16/fp16 CrossEncoder training (#3793)
Training a CrossEncoder reranker in bf16/fp16 crashed in the learning-to-rank
losses because the internal logits matrix defaults to float32 while the model
emits bf16/fp16 logits, raising "Index put requires the source and destination
dtypes match". Upcasting the logits to float32 right after the activation fixes
all six affected losses (PListMLELoss, LambdaLoss, ListNetLoss, ADRMSELoss, and
ListMLELoss/RankNetLoss via inheritance) and keeps exp/log/cumsum/softmax in a
numerically stable precision. It is a no-op for the usual float32 training.

Adds a regression test covering the six losses under bfloat16.

+++ sentence_transformers/cross_encoder/losses/adr_mse.py
@@ -178,7 +178,7 @@ class ADRMSELoss(nn.Module):
             logits_list.append(logits)
 
         logits = torch.cat(logits_list, dim=0)
-        logits = self.activation_fn(logits)
+        logits = self.activation_fn(logits).float()
 
         # Place logits into a padded matrix
         logits_matrix = torch.full((batch_size, max_docs), -1e16, device=self.model.device)
