+++ sentence_transformers/cross_encoder/losses/list_net.py
@@ -152,7 +152,7 @@ class ListNetLoss(nn.Module):
             logits_list.append(logits)
 
         logits = torch.cat(logits_list, dim=0)
-        logits = self.activation_fn(logits)
+        logits = self.activation_fn(logits).float()
 
         # Create output tensor filled with 0 (padded logits will be ignored via labels)
         logits_matrix = torch.full((batch_size, max_docs), -1e16, device=self.model.device)
